Fixing Edited Text Session State

+++ pages/TextEditor.py
@@ -39,24 +39,27 @@ st.image(r"DigiScribe_Logo.png", width = 750)
 # st.title("**:gray[Text Editor]**")
 with st.container(horizontal_alignment = "right"):
     with st.popover("Insert Text"):
-        st.write("**Current Text:**")
-        st.text_area(label = "Edit Text", value = st.session_state.edited_text, key = "edit_text")
-        if st.button("Edit Text"):
+        # st.write("**Current Text:**")
+        st.text_area(label = "**Edit Current Text**", value = st.session_state.edited_text, key = "edit_text")
+        if st.button("Edit Text:"):
             st.session_state.edited_text = st.session_state.edit_text
 
 with st.container(border = True):
     st.markdown(st.session_state.refined_text)
     st.caption("Bolded text represents segments where the model confidence was low.")
 
-content = st_quill(value = st.session_state.refined_text.replace("*", ""), html = True)
-markdown_text = markdownify.markdownify(content)
-st.session_state.edited_text = markdown_text
-
-with open("digiscribe_md.md", "w") as f:
-        f.write(markdown_text)
+content = st_quill(value = st.session_state.edited_text.replace("*", ""), html = True)
 
 col1, col2 = st.columns(2)
 dowload_cont = col1.container(horizontal_alignment = "left")
 next_cont = col2.container(horizontal_alignment = "right")
+
+confirm_button = dowload_cont.button("Confirm Text")
+if confirm_button:
+    markdown_text = markdownify.markdownify(content)
+    st.session_state.edited_text = markdown_text
+    with open("digiscribe_md.md", "w") as f:
+        f.write(markdown_text)
+
 download_md = dowload_cont.download_button("Download as Markdown", file_name = r"digiscribe_md.md", data = markdown_text, icon=":material/download:", on_click = "ignore")
 next_button = next_cont.page_link(page = r"pages/StudentHub.py",icon = ":material/school:", label = "Student Hub")
